ash_policy_access v0.1.0 AshPolicyAccess.Checker

Determines if a set of authorization requests can be met or not.

To read more about boolean satisfiability, see this page: https://en.wikipedia.org/wiki/Boolean_satisfiability_problem. At the end of the day, however, it is not necessary to understand exactly how Ash takes your authorization requirements and determines if a request is allowed. The important thing to understand is that Ash may or may not run any/all of your authorization rules as they may be deemed unnecessary. As such, authorization checks should have no side effects. Ideally, the checks built-in to ash should cover the bulk of your needs.

If you need to write your own checks see #TODO: Link to a guide about writing checks here.

Functions

find_real_scenarios(scenarios, facts)

strict_check_facts(authorizer)

strict_check_scenarios(authorizer)